requirements.txt
fastapi
uvicorn
requests
streamlit
pandas
python-dotenv
numpy